Language/requirements.txt
numpy
sounddevice
google-cloud-speech
google-cloud-texttospeech
google-cloud-translate
webrtcvad
pydub
pynput
